# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server eshop server.",
        "title": "Eshop API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/",
    "paths": {
...
        "/admin_api/v1/product_groups": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台主题产品管理"
                ],
                "summary": "主题产品列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "total",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "q[name]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "q[created_at_gteq]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "q[created_at_lteq]",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台主题产品管理"
                ],
                "summary": "添加主题产品",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ProductGroupParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
        "/admin_api/v1/product_groups/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台主题产品管理"
                ],
                "summary": "主题产品详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "productGroup id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台主题产品管理"
                ],
                "summary": "更新主题产品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "productGroup id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "productGroup params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.ProductGroupParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台主题产品管理"
                ],
                "summary": "删除主题产品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "productGroup id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
...
        "/admin_api/v1/property_names": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台规格名管理"
                ],
                "summary": "规格列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台规格名管理"
                ],
                "summary": "添加规格名",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.PropertyParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
        "/admin_api/v1/property_names/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台规格名管理"
                ],
                "summary": "规格名详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "property_name id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台规格名管理"
                ],
                "summary": "更新规格名",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "property_name params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.PropertyParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "后台规格名管理"
                ],
                "summary": "删除规格名",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "property_name id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apiHelpers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apiHelpers.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "models.JSON": {
            "type": "array",
            "items": {
                "type": "integer"
            }
        },
...
        "v1.ProductGroupParams": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "product_ids": {
                    "type": "object",
                    "$ref": "#/definitions/models.JSON"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
...
        "v1.PropertyParams": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "property_values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.PropertyValueParams"
                    }
                }
            }
        },
        "v1.PropertyValueParams": {
            "type": "object",
            "required": [
                "property_name_id",
                "value"
            ],
            "properties": {
                "_destroy": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "property_name_id": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}